+++ src/dsl/parser/expr.rs
@@ -356,6 +356,8 @@ impl Parser {
             | Some(Token::Sum)
             | Some(Token::Mean)
             | Some(Token::Stdev)
+            | Some(Token::Fft)
+            | Some(Token::Ifft)
             | Some(Token::Scale)
             | Some(Token::Reshape)
             | Some(Token::Stack) => return self.parse_call_expr(),
@@ -573,6 +575,8 @@ impl Parser {
             Some(Token::Sum) => CallExpr::Sum(Box::new(self.parse_simple_expr()?)),
             Some(Token::Mean) => CallExpr::Mean(Box::new(self.parse_simple_expr()?)),
             Some(Token::Stdev) => CallExpr::Stdev(Box::new(self.parse_simple_expr()?)),
+            Some(Token::Fft) => CallExpr::Fft(Box::new(self.parse_simple_expr()?)),
+            Some(Token::Ifft) => CallExpr::Ifft(Box::new(self.parse_simple_expr()?)),
             Some(Token::Scale) => {
                 let input = self.parse_simple_expr()?;
                 self.eat(&Token::By)?;